backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
bcrypt==4.1.2
PyMySQL==1.1.0
python-dotenv==1.0.0
stripe==7.14.0
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
flask-socketio==5.3.6
eventlet==0.35.2
gunicorn==21.2.0
